requirements.txt
google-generativeai==0.3.2
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.24.3 